+++ src/main/java/com/github/flooooooooooorian/meinkochbuch/dtos/cookbook/CookbookPreview.java
@@ -7,8 +7,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -22,5 +20,5 @@ public class CookbookPreview {
 
     private ImageDto thumbnail;
 
-    private BigDecimal ratingAverage;
+    private double ratingAverage;
 }
